Revert "Moves Pebble plan types and helpers to public. (#152)"
This reverts commit fb9b611ab1159c31348a21594809bc19d0ddd62a.
